setup AttemptForgingKeyView to simplify forging future processing #1108
