+++ src/test/scala/co/topl/api/RPCMockState.scala
@@ -7,8 +7,7 @@ import akka.http.scaladsl.server.Route
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import akka.pattern.ask
 import akka.util.{ByteString, Timeout}
-import co.topl.attestation.keyManagement.{KeyManager, KeyManagerRef}
-import co.topl.consensus.{Forger, ForgerRef}
+import co.topl.consensus.{Forger, ForgerRef, KeyManager, KeyManagerRef}
 import co.topl.http.HttpService
 import co.topl.http.api.ApiEndpoint
 import co.topl.http.api.endpoints._
